feature: cria ecossistema de configurações

+++ app/Http/Controllers/Admin/AdminController.php
@@ -7,21 +7,26 @@ use App\Http\Requests\Admin\StoreAdminRequest;
 use App\Http\Requests\Admin\UpdateAdminRequest;
 use App\Http\Resources\Admin\AdminResource;
 use App\Models\Admin;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Http\JsonResponse;
 
 class AdminController extends Controller
 {
     public function index()
     {
-        return AdminResource::collection(Admin::paginate());
+        return AdminResource::collection(
+            Admin::query()
+                ->latest()
+                ->paginate(20)
+        );
     }
 
     public function store(StoreAdminRequest $request)
     {
-        $data = $request->validated();
-        $data['password'] = Hash::make($data['password']);
-        $admin = Admin::create($data);
-        return AdminResource::make($admin)->response()->setStatusCode(201);
+        $admin = Admin::create($request->validated());
+
+        return AdminResource::make($admin)
+            ->response()
+            ->setStatusCode(201);
     }
 
     public function show(Admin $admin)
@@ -32,18 +37,27 @@ class AdminController extends Controller
     public function update(UpdateAdminRequest $request, Admin $admin)
     {
         $data = $request->validated();
-        if (! empty($data['password'])) {
-            $data['password'] = Hash::make($data['password']);
-        } else {
+
+        if (empty($data['password'])) {
             unset($data['password']);
         }
+
         $admin->update($data);
+
         return AdminResource::make($admin);
     }
 
-    public function destroy(Admin $admin)
+    public function destroy(Admin $admin): JsonResponse
     {
+        if (auth('admin')->id() === $admin->id) {
+            return response()->json([
+                'message' => 'Você não pode remover o próprio usuário logado.',
+            ], 422);
+        }
+
         $admin->delete();
+
         return response()->json(null, 204);
     }
 }
+
